[trace-guest] Avoid deadlock when calling `outb` in exceptions context to report trace data
When guest tracing is enabled and an exception occurs while the guest
trace state lock is acquired (e.g. no more memory left to allocate a
trace event on the heap), the guest tries to report to the host what the
error was by calling the `out32` function.
This function reports the guest trace data by serializing it on the spot,
which needs to acquire the lock again, this leads to a deadlock.

Signed-off-by: Doru Blânzeanu <dblnz@pm.me>